backend/requirements.txt
fastapi==0.120.0
uvicorn==0.38.0
sqlalchemy==2.0.44
alembic==1.17.0
python-multipart==0.0.20
websockets==15.0.1
python-dotenv==1.2.1
